requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Werkzeug==2.2.3
python-dotenv==1.0.0
email-validator==2.0.0